tests: Migrate MsgInfoView tests to test_popups.
